+++ threedi_models_and_simulations/data_models/simulation_data_models.py
@@ -96,6 +96,7 @@ class InitialConditions(SimulationElement):
     aggregation_method_groundwater: str = None
     saved_state: str = None
     initial_concentrations_2d: dict = None
+    initial_concentrations_1d: dict = None
 
 
 @dataclass
@@ -146,6 +147,7 @@ class Precipitation(SimulationElement):
     netcdf_filepath: str = None
     netcdf_global: bool = None
     netcdf_raster: bool = None
+    substances: list = None
 
 
 @dataclass