+++ HISTORY.rst
@@ -4,13 +4,19 @@ History
 3.14.2 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Fixes/enhancements: #616
+- Setting 1D initial substance concentration (#609)
+- Added check for non-ascii characters in substance unit (#621)
+- Added user feedback for invalid boundary conditions file. (#624)
+- Added substance concentration section in Precipation page. (#537)
+- Added diffusion parameter to Substances page. (#602)
 
 
 3.14.1 (2024-09-25)
 -------------------
 
 - Fix for the 'simulation_user_first_name' KeyError.
+- Add new 1D advection options
 
 
 3.14.0 (2024-09-24)
@@ -26,6 +32,9 @@ History
 - Fix error when trying to start a simulation with laterals (#589).
 - Updated pyqtgraph to version 0.13.7
 - Model selection: fixed bug (numeric and date columns were sorted lexicographically) (#587)
+- Simulation template creation: append prefix to saved template name in case of Multiple Simulations (#613)
+- GUI: Load parameters from template: not loading substance decay coefficient (#612)
+- GUI: Add substance concentrations section to Precipitation page (#537)
 
 
 3.12.0 (2024-07-17)